1.ICyclingDataUtil.cs脚本，NetOriginData类的duration字段的类型改为long。int型太短，会报错。 2.GameStart.prefab预制体，版本提升至1.0.17。

+++ Assets/Scripts/Data/Remote/Interface/ICyclingDataUtil.cs
@@ -52,7 +52,7 @@ namespace AppGame.Data.Remote
     {
         public string child_sn { get; set; }
         public int? distance { get; set; }
-        public int? duration { get; set; }
+        public long? duration { get; set; }
         public int? calories { get; set; }
 
         public OriginData ToOriginData()
